Exercise with Java
 ControlFlowExercises.java:

 Added:
  - Made a last minute change to one of the Do While loops. I was using
  an integer when I should have been using a long to represent such
  a large number.

+++ src/io/ControlFlowExercises.java
@@ -29,7 +29,7 @@ public class ControlFlowExercises {
             a -= 5;
         } while (a >= -10);
 
-        int b = 2;
+        long b = 2;
         do {
             System.out.println(b);
             b *= b;
